@@ -1,20 +1,21 @@
 package com.muhardin.endy.belajar.spring.security.multiauth.config;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.web.configurers.oauth2.server.resource.OAuth2ResourceServerConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
 
-@Configuration @Order(1)
-public class ResourceServerConfig extends WebSecurityConfigurerAdapter {
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.antMatcher("/api/**")
-                .authorizeRequests(authorizeRequests ->
-                        authorizeRequests
-                                .anyRequest().authenticated())
-                //.oauth2ResourceServer(OAuth2ResourceServerConfigurer::opaqueToken);
-                .oauth2ResourceServer(OAuth2ResourceServerConfigurer::jwt);
+@Configuration
+public class ResourceServerConfig {
+
+    @Bean @Order(1)
+    public SecurityFilterChain apiSecurity(HttpSecurity http) throws Exception {
+        http.securityMatcher("/api/**")
+        .authorizeHttpRequests(auth -> auth
+            .anyRequest().authenticated()
+        ).oauth2ResourceServer(OAuth2ResourceServerConfigurer::jwt);
+        return http.build();
     }
 }
